26022016 result dump fix / del virtualenv

+++ tools/analyzer.py
@@ -22,11 +22,23 @@ if __name__=="__main__":
 	Analyser.DF=np.float(args.df)
 	Analyser.result_path=args.r
 
+	results={}
+	results['RSID']=np.array([])
+	results['p_value']=np.array([])
+	results['t-stat']=np.array([])
+	results['phenotype']=np.array([])
+	results['SE']=np.array([])
+
 	while True:
 		Analyser.summary()
 		if Analyser.results is None:
 			break
 		print('Saving data...')
+		if not os.path.exists(os.path.join(args.out,'results'+'.csv')):
+			df=pd.DataFrame.from_dict(results)
+			df.to_csv( os.path.join(args.out,'results'+'.csv'), sep=" "  )
 		df=pd.DataFrame.from_dict(Analyser.results)
-		df.to_csv( os.path.join(args.out,'results'+'.csv'), sep=" "  )
+		with open(os.path.join(args.out,'results'+'.csv'), 'a') as f:
+			df.to_csv(f, sep=" ",header=False)
+
